refactor(provenance): clean up skillbook structure and insight sources
- Remove TraceReference and trace traversal engine (~250 lines)
- Slim InsightSource to pure identity: 8 fields (was 14)
- Remove tagging system (helpful/harmful/neutral counters, TagStep, tag_skill)
- Consolidate ace/insight_source.py into ace/steps/attach_insight_sources.py
- Replace fuzzy batch matching with deterministic matching
- Fix Skillbook.to_dict() to use InsightSource.to_dict() (sparse output)
- Omit embedding:null from serialization
- Add thread safety to source_map/summary/filter
- Document provider SDK requirements in pydantic_ai.py and migration doc
