+++ tests/test_pipeline_exports.py
@@ -115,7 +115,6 @@ class TestStepExports:
             "AgentStep",
             "EvaluateStep",
             "ReflectStep",
-            "TagStep",
             "UpdateStep",
             "AttachInsightSourcesStep",
             "ApplyStep",
@@ -140,7 +139,6 @@ class TestStepExports:
             "AgentStep",
             "EvaluateStep",
             "ReflectStep",
-            "TagStep",
             "UpdateStep",
             "AttachInsightSourcesStep",
             "ApplyStep",
@@ -172,7 +170,7 @@ class TestBuildSteps:
             skill_manager=MockSkillManager(),
         )
         assert isinstance(steps, list)
-        assert len(steps) >= 5  # Agent, Evaluate, Reflect, Tag, Update, Sources, Apply
+        assert len(steps) >= 4  # Agent, Evaluate, Reflect, Update, Sources, Apply
         assert isinstance(steps[0], AgentStep)
         assert isinstance(steps[1], EvaluateStep)
         assert isinstance(steps[2], ReflectStep)
@@ -186,7 +184,7 @@ class TestBuildSteps:
             skill_manager=MockSkillManager(),
         )
         assert isinstance(steps, list)
-        assert len(steps) >= 5  # Reflect, Tag, Update, Sources, Apply
+        assert len(steps) >= 4  # Reflect, Update, Sources, Apply
         assert isinstance(steps[0], ReflectStep)
 
     def test_ace_from_roles_delegates_to_build_steps(self):
